requirements.txt
pydantic
abnf
python-dateutil
ecdsa